fix discourse category urls, missing utils bug.
